+++ site/dat/docs/Changelog.md
@@ -2,6 +2,7 @@
 
 ## 0.4.4 (next)
  - limit max size of file that goes into zip for artifact upload (max-size option)
+ - fix cumulative KPIset recalculating
 
 ## 0.4.3
  - allow passing config as stdin to CLI like `./generate-config.sh | bzt`